add .gitignore and fix ft_putchar_fd.c is an int function

+++ ft_putchar_fd.c
@@ -6,13 +6,13 @@
 /*   By: faguirre <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/18 01:25:38 by faguirre          #+#    #+#             */
-/*   Updated: 2024/09/18 02:01:57 by faguirre         ###   ########.fr       */
+/*   Updated: 2024/10/12 10:41:59 by faguirre         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <unistd.h>
 
-ssize_t	ft_putchar_fd(char c, int fd)
+int	ft_putchar_fd(char c, int fd)
 {
 	if (write(fd, &c, 1) == -1)
 		return (0);
